- add title - fix #157 - refactoring some ui - fix system info statics view when system node having no effect - fix note when empty (auto delete)
