+++ src/WHMapper/Pages/Mapper/SystemInfos/Overview.cs
@@ -15,13 +15,7 @@ namespace WHMapper.Pages.Mapper.SystemInfos
 {
     [Authorize(Policy = "Access")]
     public partial class Overview : ComponentBase
-    {
-        private const string MSG_SOLAR_SYSTEM_COMMENT_AUTOSAVE_SUCCESS = "Solar system comment autosave successfull";
-        private const string MSG_SOLAR_SYSTEM_COMMENT_AUTOSAVE_ERROR = "Solar system comment autosave error";
-        private const string MSG_SOLAR_SYSTEM_COMMENT_AUTOUPDATE_SUCCESS = "Solar system comment autoupdate successfull";
-        private const string MSG_SOLAR_SYSTEM_COMMENT_AUTOUPDATE_ERROR = "Solar system comment autoupdate error";
-        private const string MSG_AUTOSAVE_OR_UPDATE_ERROR = "Solar system autosave or update error";
-
+    {        
         private const string NO_EFFECT = "No Effect";
 
         private string _secColor = string.Empty;
@@ -30,11 +24,6 @@ namespace WHMapper.Pages.Mapper.SystemInfos
 
         private string _systemType = string.Empty;
         private string _effect = NO_EFFECT;
-
-        private string _solarSystemComment = string.Empty;
-        private WHNote? _note = null!;
-
-
         [Inject]
         private IWHNoteRepository DbWHNotes { get; set; } = null!;
 
@@ -57,7 +46,6 @@ namespace WHMapper.Pages.Mapper.SystemInfos
 
             if (CurrentSystemNode != null)
             {
-                _solarSystemComment = string.Empty;
                 _secColor = WHColorHelper.GetSecurityStatusColor(CurrentSystemNode.SecurityStatus);
                 _systemColor = WHColorHelper.GetSystemTypeColor(CurrentSystemNode.SystemType);
                 _whEffectColor = WHColorHelper.GetEffectColor(CurrentSystemNode.Effect);
@@ -80,18 +68,6 @@ namespace WHMapper.Pages.Mapper.SystemInfos
                     _effect = NO_EFFECT;
                 else
                     _effect = GetWHEffectValueAsString(CurrentSystemNode.Effect);
-
-
-                //load system saved notes
-                if (DbWHNotes != null)
-                {
-                    _note = await DbWHNotes.GetBySolarSystemId(CurrentSystemNode.SolarSystemId);
-                    if (_note == null)
-                        _solarSystemComment = string.Empty;
-                    else
-                        _solarSystemComment = _note.Comment;
-                }
-
             }
         }
 
@@ -109,99 +85,6 @@ namespace WHMapper.Pages.Mapper.SystemInfos
                 return effect.ToString();
             }
         }
-
-        private async Task OnNoteChanged()
-        {
-            if (_timer == null && !string.IsNullOrEmpty(_solarSystemComment))
-            {
-                Task.Run(() => HandleTimerAsync());
-            }
-        }
-
-        private PeriodicTimer _timer = null!;
-        private CancellationTokenSource _cts = null!;
-
-        private string _previousValue;
-
-        private async Task HandleTimerAsync()
-        {
-
-            if (_timer == null && !string.IsNullOrEmpty(_solarSystemComment))
-            {
-                _previousValue = _solarSystemComment;
-                _cts = new CancellationTokenSource();
-                _timer = new PeriodicTimer(TimeSpan.FromMilliseconds(500));
-                try
-                {
-                    while (await _timer.WaitForNextTickAsync(_cts.Token))
-                    {
-                        if(_previousValue== _solarSystemComment)
-                        {
-                            if (_note == null)
-                            {
-                                try
-                                {
-                                    _note = await DbWHNotes.Create(new WHNote(CurrentSystemNode.SolarSystemId, _solarSystemComment));
-                                    if (_note != null)
-                                        Snackbar.Add(MSG_SOLAR_SYSTEM_COMMENT_AUTOSAVE_SUCCESS, Severity.Success);
-                                    else
-                                    {
-                                        Logger.LogError(MSG_SOLAR_SYSTEM_COMMENT_AUTOUPDATE_ERROR);
-                                        Snackbar.Add(MSG_SOLAR_SYSTEM_COMMENT_AUTOSAVE_ERROR, Severity.Error);
-                                    }
-                                }
-                                catch(Exception autoSaveEx)
-                                {
-
-                                    Snackbar.Add(MSG_SOLAR_SYSTEM_COMMENT_AUTOSAVE_ERROR, Severity.Error);
-                                    
-                                }
-                            }
-                            else
-                            {
-                                try
-                                {
-                                    _note.Comment = _solarSystemComment;
-                                    _note = await DbWHNotes.Update(_note.Id, _note);
-                                    if (_note != null)
-                                        Snackbar.Add(MSG_SOLAR_SYSTEM_COMMENT_AUTOUPDATE_SUCCESS, Severity.Success);
-                                    else
-                                    {
-                                        Logger.LogError(MSG_SOLAR_SYSTEM_COMMENT_AUTOUPDATE_ERROR);
-                                        Snackbar.Add(MSG_SOLAR_SYSTEM_COMMENT_AUTOUPDATE_ERROR, Severity.Error);
-                                    }
-                                }
-                                catch (Exception exAutoSave)
-                                {
-                                    Logger.LogError(MSG_SOLAR_SYSTEM_COMMENT_AUTOUPDATE_ERROR, exAutoSave);
-                                    Snackbar.Add(MSG_SOLAR_SYSTEM_COMMENT_AUTOUPDATE_ERROR, Severity.Error);
-                                }
-                            }
-
-                            _cts.Cancel();
-                        }
-                        else
-                        {
-                            _previousValue = _solarSystemComment;
-                        }
-                    }
-                }
-                catch(OperationCanceledException)
-                {
-                }
-                catch (Exception ex)
-                {
-                    Logger.LogError(MSG_AUTOSAVE_OR_UPDATE_ERROR, ex);
-                }
-                finally
-                {
-                    _timer = null!;
-                    _cts = null!;
-                    _previousValue = string.Empty;
-                }
-            }
-        }
-
     }
 }
