+++ src/WHMapper/Pages/_Layout.cshtml
@@ -12,11 +12,12 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
+    <title>Eve WHMapper</title>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <base href="~/" />
     <link rel="stylesheet" href="css/bootstrap/bootstrap.min.css" />
-    <!-- <link href="css/site.css" rel="stylesheet" />-->
+
 
     <link href="WHMapper.styles.css" rel="stylesheet" />
     <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" rel="stylesheet" />
